Solve issue : one security config wasn't used.
